requirements.txt
Pillow>=9.0
bleak>=0.22
crc8>=0.2.0
pyserial>=3.5
